+++ tests/mocks/reacti18nMock.ts
@@ -1,0 +1,15 @@
+// Mock react-i18next
+jest.mock('react-i18next', () => ({
+  useTranslation: () => {
+    return {
+      t: (str: string) => str,
+      i18n: {
+        changeLanguage: () => new Promise(() => {}),
+      },
+    };
+  },
+  withTranslation: () => (Component: { defaultProps: any; }) => {
+    Component.defaultProps = { ...Component.defaultProps, t: (str: string) => str };
+    return Component;
+  },
+}));